pyproject.toml
[project]
name = "gsp-render-dev"
version = "0.1.0"
description = "Unified MotrixSim 3D Gaussian rendering demos and benchmark notebooks"
readme = "README.md"
requires-python = ">=3.10,<3.11"
dependencies = [
    "motrixsim_core==0.7.1.dev97295",
    "torch==2.4.1+cu121",
    "triton==3.0.0",
    "gaussian_renderer==0.2.0",
    "gsplat==1.5.3",
    "numpy==2.2.6",
    "scipy==1.15.3",
    "plyfile==1.1.3",
    "mediapy==1.2.6",
    "matplotlib==3.10.9",
    "ipykernel==7.2.0",
    "nbconvert>=7.16",
    "onnxruntime==1.22.1",
    "packaging==25.0",
    "setuptools==80.9.0",
    "wheel==0.45.1",
    "ninja==1.13.0",
    "rich==15.0.0",
]

[[tool.uv.index]]
name = "pypi-tuna"
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
default = true

[[tool.uv.index]]
name = "motphys-pypi"
url = "https://pypi.motphys.com/simple/"
explicit = true

[[tool.uv.index]]
name = "pytorch-cu121"
url = "https://download.pytorch.org/whl/cu121"
explicit = true

[tool.uv.sources]
motrixsim_core = { index = "motphys-pypi" }
torch = [
    { index = "pytorch-cu121", marker = "sys_platform == 'linux' and platform_machine == 'x86_64'" },
]

[tool.uv]
package = false
prerelease = "allow"
environments = [
    "sys_platform == 'linux' and platform_machine == 'x86_64'",
]
